requirements.txt
backports.lzma
requests
frida
android sdk with path variable pointing to build tools and platform tools
android sdk version >=4.2
DroidBot installed to path
Python == 2.7
JDK >= 1.7
androguard